Fixed spawning and added additional configuration options.

+++ common/src/main/java/net/creeperhost/chickens/containers/BreederMenu.java
@@ -6,6 +6,7 @@ import net.creeperhost.polylib.client.modulargui.lib.container.DataSync;
 import net.creeperhost.polylib.client.modulargui.lib.container.SlotGroup;
 import net.creeperhost.polylib.containers.PolyBlockContainerMenu;
 import net.creeperhost.polylib.containers.slots.PolySlot;
+import net.creeperhost.polylib.data.serializable.FloatData;
 import net.creeperhost.polylib.data.serializable.IntData;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Inventory;
@@ -18,7 +19,7 @@ public class BreederMenu extends PolyBlockContainerMenu<BreederBlockEntity> {
     public final SlotGroup chickens = createSlotGroup(2, 0);
     public final SlotGroup output = createSlotGroup(3, 0);
 
-    public final DataSync<Integer> progress;
+    public final DataSync<Float> progress;
 
     public BreederMenu(int windowId, Inventory playerInv, FriendlyByteBuf extraData) {
         this(windowId, playerInv, getClientTile(playerInv, extraData));
@@ -26,7 +27,7 @@ public class BreederMenu extends PolyBlockContainerMenu<BreederBlockEntity> {
 
     public BreederMenu(int windowId, Inventory playerInv, BreederBlockEntity tile) {
         super(ModContainers.BREEDER_CONTAINER.get(), windowId, playerInv, tile);
-        progress = new DataSync<>(this, new IntData(), tile.progress::get);
+        progress = new DataSync<>(this, new FloatData(), tile.progress::get);
 
         main.addPlayerMain(playerInv);
         hotBar.addPlayerBar(playerInv);
